refactor(payments): Removed payments from clickhouse
