+++ src/interface/event/Event.ts
@@ -19,19 +19,13 @@ export type AITokenUsageEventData = {
   metadata?: Record<string, unknown>;
 };
 
-export type PaymentEventData = {
-  creditAmount: number;
-};
-
 export type EventKind =
   | "BASIC_USAGE"
   | "AI_TOKEN_USAGE"
-  | "PAYMENT"
 
 export type EventDataMap = {
   BASIC_USAGE: BasicUsageEventData;
   AI_TOKEN_USAGE: AITokenUsageEventData;
-  PAYMENT: PaymentEventData;
 };
 
 export type EventData<K extends EventKind> = EventDataMap[K];
@@ -39,7 +33,6 @@ export type EventData<K extends EventKind> = EventDataMap[K];
 export type SqlRecord =
   | { type: "BASIC_USAGE"; reported_timestamp: DateTime; data: BasicUsageEventData; userId: UserId }
   | { type: "AI_TOKEN_USAGE"; reported_timestamp: DateTime; data: AITokenUsageEventData; userId: UserId }
-  | { type: "PAYMENT"; reported_timestamp: DateTime; data: PaymentEventData; userId: UserId }
 
 export type SqlRecordOf<K extends EventKind> = Extract<SqlRecord, { type: K }>;
 
@@ -65,6 +58,4 @@ export interface AITokenUsageEvent extends Event<"AI_TOKEN_USAGE"> {
   readonly reportedTimestamp: DateTime;
 }
 
-export interface PaymentEvent extends Event<"PAYMENT"> {
-  readonly userId: UserId;
-}
+
