+++ src/storage/adapter/clickhouse/handlers/index.ts
@@ -1,6 +1,5 @@
 export { handleAddBasicUsage } from "./addBasicUsage";
 export { handleAddAiTokenUsage } from "./addAiTokenUsage";
-export { handleAddPayment } from "./addPayment";
 export { handlePriceRequestBasicUsage } from "./priceRequestBasicUsage";
 export { handlePriceRequestAiTokenUsage } from "./priceRequestAiTokenUsage";
 export { handleQueryEvents } from "./queryEvents";
