+++ src/storage/adapter/clickhouse/handlers/queryEvents.ts
@@ -106,33 +106,7 @@ const CH_FIELDS: Record<EventTableName, Record<ChFieldKey, ChFieldDef>> = {
     },
     creditAmount: { select: null, where: null },
     provider: { select: "provider", where: "provider" },
-    metadata:          { select: "toString(metadata)",                      where: null },
-  },
-  payment_events: {
-    eventId: { select: "toString(id)", where: "id" },
-    eventType: { select: "'PAYMENT'", where: null },
-    userId: { select: "user_id", where: "user_id" },
-    apiKeyId: { select: "api_key_id", where: "api_key_id" },
-    reportedTimestamp: {
-      select: "toString(reported_timestamp)",
-      where: "reported_timestamp",
-    },
-    ingestedTimestamp: {
-      select: "toString(ingested_timestamp)",
-      where: "ingested_timestamp",
-    },
-    basicUsageType: { select: null, where: null },
-    debitAmount: { select: null, where: null },
-    model: { select: null, where: null },
-    inputTokens: { select: null, where: null },
-    outputTokens: { select: null, where: null },
-    inputDebitAmount: { select: null, where: null },
-    outputDebitAmount: { select: null, where: null },
-    inputCacheTokens: { select: null, where: null },
-    inputCacheDebitAmount: { select: null, where: null },
-    creditAmount: { select: "toString(credit_amount)", where: "credit_amount" },
-    provider: { select: null, where: null },
-    metadata: { select: null, where: null },
+    metadata: { select: "toString(metadata)", where: null },
   },
 };
