+++ src/storage/adapter/postgres/handlers/index.ts
@@ -1,5 +1,4 @@
 export { handleAddBasicUsage } from "./addBasicUsage";
-export { handleAddPayment } from "./addPayment";
 export { handlePriceRequestBasicUsage } from "./priceRequestBasicUsage";
 export { handleAddAiTokenUsage } from "./addAiTokenUsage";
 export { handlePriceRequestAiTokenUsage } from "./priceRequestAiTokenUsage";
