+++ src/storage/adapter/postgres/handlers/queryEvents.ts
@@ -129,42 +129,6 @@ const PG_FIELDS: PGFieldRegistry = {
     provider: { select: "provider", whereCol: "provider", whereCast: "" },
     metadata: { select: "metadata::text", whereCol: null, whereCast: "" },
   },
-  payment_events: {
-    eventId: { select: "id::text", whereCol: "id", whereCast: "::uuid" },
-    eventType: { select: "'PAYMENT'", whereCol: null, whereCast: "" },
-    userId: { select: "user_id::text", whereCol: "user_id", whereCast: "" },
-    apiKeyId: {
-      select: "api_key_id::text",
-      whereCol: "api_key_id",
-      whereCast: "",
-    },
-    reportedTimestamp: {
-      select: "reported_timestamp::text",
-      whereCol: "reported_timestamp",
-      whereCast: "::timestamptz",
-    },
-    ingestedTimestamp: {
-      select: "ingested_timestamp::text",
-      whereCol: "ingested_timestamp",
-      whereCast: "::timestamptz",
-    },
-    basicUsageType: { select: null, whereCol: null, whereCast: "" },
-    debitAmount: { select: null, whereCol: null, whereCast: "" },
-    model: { select: null, whereCol: null, whereCast: "" },
-    inputTokens: { select: null, whereCol: null, whereCast: "" },
-    outputTokens: { select: null, whereCol: null, whereCast: "" },
-    inputDebitAmount: { select: null, whereCol: null, whereCast: "" },
-    outputDebitAmount: { select: null, whereCol: null, whereCast: "" },
-    inputCacheTokens: { select: null, whereCol: null, whereCast: "" },
-    inputCacheDebitAmount: { select: null, whereCol: null, whereCast: "" },
-    creditAmount: {
-      select: "credit_amount::text",
-      whereCol: "credit_amount",
-      whereCast: "::bigint",
-    },
-    provider: { select: null, whereCol: null, whereCast: "" },
-    metadata: { select: null, whereCol: null, whereCast: "" },
-  },
 };
 
 const OUTPUT_FIELDS = Object.keys(PG_FIELDS.basic_usage_events);
@@ -311,7 +275,9 @@ async function handleAggregationQuery(
       if (def?.aggExpr) {
         cols.push(sql`${sql.raw(def.aggExpr)} as ${sql.raw(`"agg_value"`)}`);
       } else if (def?.whereCol) {
-        cols.push(sql`${sql.raw(def.whereCol)}::bigint as ${sql.raw(`"agg_value"`)}`);
+        cols.push(
+          sql`${sql.raw(def.whereCol)}::bigint as ${sql.raw(`"agg_value"`)}`
+        );
       } else {
         cols.push(sql`0::bigint as ${sql.raw(`"agg_value"`)}`);
       }
